Implements level-driven constraint for DPB size and refactors level.c implementation
- Addresses TODO comments about DPB size constraint

- Moves level.h and level.c files to from av2/encoder/ to av2/common/

- Creates a basis for other constraints referred in #1188

Partial fix for #1188 (only for DPB size constraints)
