+++ av2/av2.cmake
@@ -57,6 +57,8 @@ list(
   "${AVM_ROOT}/av2/common/frame_buffers.h"
   "${AVM_ROOT}/av2/common/idct.c"
   "${AVM_ROOT}/av2/common/idct.h"
+  "${AVM_ROOT}/av2/common/level.c"
+  "${AVM_ROOT}/av2/common/level.h"
   "${AVM_ROOT}/av2/common/mv.h"
   "${AVM_ROOT}/av2/common/mvref_common.c"
   "${AVM_ROOT}/av2/common/mvref_common.h"
@@ -227,8 +229,6 @@ list(
   "${AVM_ROOT}/av2/encoder/hybrid_fwd_txfm.h"
   "${AVM_ROOT}/av2/encoder/interp_search.c"
   "${AVM_ROOT}/av2/encoder/interp_search.h"
-  "${AVM_ROOT}/av2/encoder/level.c"
-  "${AVM_ROOT}/av2/encoder/level.h"
   "${AVM_ROOT}/av2/encoder/lookahead.c"
   "${AVM_ROOT}/av2/encoder/lookahead.h"
   "${AVM_ROOT}/av2/encoder/mcomp.c"
