+++ av2/encoder/encoder.h
@@ -29,6 +29,7 @@
 #include "av2/common/bru.h"
 #include "av2/common/entropymode.h"
 #include "av2/common/enums.h"
+#include "av2/common/level.h"
 #include "av2/common/pred_common.h"
 #include "av2/common/resize.h"
 #include "av2/common/thread_common.h"
@@ -41,7 +42,6 @@
 #include "av2/encoder/encodemb.h"
 #include "av2/encoder/firstpass.h"
 #include "av2/encoder/global_motion.h"
-#include "av2/encoder/level.h"
 #include "av2/encoder/lookahead.h"
 #include "av2/encoder/mcomp.h"
 #include "av2/encoder/ratectrl.h"
